app/config.py
import os
import redis
from dotenv import load_dotenv

load_dotenv()

OPENAI_API_KEY = os.getenv("OPENAI_API_KEY")
REDIS_URL = os.getenv("REDIS_URL")
DATABASE_URL = os.getenv("DATABASE_URL")

redis_client = redis.StrictRedis.from_url(REDIS_URL, decode_responses=True)
