requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
asyncpg
pydantic
redis
aioredis
celery
aiohttp
python-dotenv
pytest
prometheus-client
psycopg2-binary
python-multipart
openai==1.2.3

